Fold S.S.C.Encodings into unified S.S.Cryptography
