+++ src/libraries/System.Security.Cryptography.Encoding/System.Security.Cryptography.Encoding.sln
@@ -13,10 +13,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Security.Cryptograph
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Security.Cryptography.Encoding", "src\System.Security.Cryptography.Encoding.csproj", "{4B1D276E-6576-4F42-9F38-C91C95437F20}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Security.Cryptography.Encoding.Tests", "tests\System.Security.Cryptography.Encoding.Tests.csproj", "{8A8D5FD7-557E-42EF-AA8A-5453E82574F2}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{179D2235-ABD5-45D9-8DAB-E24763305362}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ref", "ref", "{34B969D7-6BED-4597-8C25-502A29CF0518}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0806AC6F-4837-437A-A32F-EED211934481}"
@@ -55,17 +51,11 @@ Global
 		{4B1D276E-6576-4F42-9F38-C91C95437F20}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4B1D276E-6576-4F42-9F38-C91C95437F20}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4B1D276E-6576-4F42-9F38-C91C95437F20}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8A8D5FD7-557E-42EF-AA8A-5453E82574F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8A8D5FD7-557E-42EF-AA8A-5453E82574F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8A8D5FD7-557E-42EF-AA8A-5453E82574F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8A8D5FD7-557E-42EF-AA8A-5453E82574F2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{201E9083-CCA0-4AA8-A2B2-B37AC13481E3} = {179D2235-ABD5-45D9-8DAB-E24763305362}
-		{8A8D5FD7-557E-42EF-AA8A-5453E82574F2} = {179D2235-ABD5-45D9-8DAB-E24763305362}
 		{A428E072-E4DD-4F64-AFE3-DC9FC8D5563D} = {34B969D7-6BED-4597-8C25-502A29CF0518}
 		{465CB04B-D2D8-4E64-BA6D-571BE95391B4} = {34B969D7-6BED-4597-8C25-502A29CF0518}
 		{43B33B0D-4009-4838-B9F4-E5308C1827F7} = {0806AC6F-4837-437A-A32F-EED211934481}
