--- src/libraries/System.Security.Cryptography.Encoding/ref/System.Security.Cryptography.Encoding.cs
@@ -1,145 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the https://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-namespace System.Security.Cryptography
-{
-    public partial class AsnEncodedData
-    {
-        protected AsnEncodedData() { }
-        public AsnEncodedData(byte[] rawData) { }
-        public AsnEncodedData(System.ReadOnlySpan<byte> rawData) { }
-        public AsnEncodedData(System.Security.Cryptography.AsnEncodedData asnEncodedData) { }
-        public AsnEncodedData(System.Security.Cryptography.Oid? oid, byte[] rawData) { }
-        public AsnEncodedData(System.Security.Cryptography.Oid? oid, System.ReadOnlySpan<byte> rawData) { }
-        public AsnEncodedData(string oid, byte[] rawData) { }
-        public AsnEncodedData(string oid, System.ReadOnlySpan<byte> rawData) { }
-        public System.Security.Cryptography.Oid? Oid { get { throw null; } set { } }
-        public byte[] RawData { get { throw null; } set { } }
-        public virtual void CopyFrom(System.Security.Cryptography.AsnEncodedData asnEncodedData) { }
-        public virtual string Format(bool multiLine) { throw null; }
-    }
-    public sealed partial class AsnEncodedDataCollection : System.Collections.ICollection, System.Collections.IEnumerable
-    {
-        public AsnEncodedDataCollection() { }
-        public AsnEncodedDataCollection(System.Security.Cryptography.AsnEncodedData asnEncodedData) { }
-        public int Count { get { throw null; } }
-        public bool IsSynchronized { get { throw null; } }
-        public System.Security.Cryptography.AsnEncodedData this[int index] { get { throw null; } }
-        public object SyncRoot { get { throw null; } }
-        public int Add(System.Security.Cryptography.AsnEncodedData asnEncodedData) { throw null; }
-        public void CopyTo(System.Security.Cryptography.AsnEncodedData[] array, int index) { }
-        public System.Security.Cryptography.AsnEncodedDataEnumerator GetEnumerator() { throw null; }
-        public void Remove(System.Security.Cryptography.AsnEncodedData asnEncodedData) { }
-        void System.Collections.ICollection.CopyTo(System.Array array, int index) { }
-        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() { throw null; }
-    }
-    public sealed partial class AsnEncodedDataEnumerator : System.Collections.IEnumerator
-    {
-        internal AsnEncodedDataEnumerator() { }
-        public System.Security.Cryptography.AsnEncodedData Current { get { throw null; } }
-        object System.Collections.IEnumerator.Current { get { throw null; } }
-        public bool MoveNext() { throw null; }
-        public void Reset() { }
-    }
-    public partial class FromBase64Transform : System.IDisposable, System.Security.Cryptography.ICryptoTransform
-    {
-        public FromBase64Transform() { }
-        public FromBase64Transform(System.Security.Cryptography.FromBase64TransformMode whitespaces) { }
-        public virtual bool CanReuseTransform { get { throw null; } }
-        public bool CanTransformMultipleBlocks { get { throw null; } }
-        public int InputBlockSize { get { throw null; } }
-        public int OutputBlockSize { get { throw null; } }
-        public void Clear() { }
-        public void Dispose() { }
-        protected virtual void Dispose(bool disposing) { }
-        ~FromBase64Transform() { }
-        public int TransformBlock(byte[] inputBuffer, int inputOffset, int inputCount, byte[] outputBuffer, int outputOffset) { throw null; }
-        public byte[] TransformFinalBlock(byte[] inputBuffer, int inputOffset, int inputCount) { throw null; }
-    }
-    public enum FromBase64TransformMode
-    {
-        IgnoreWhiteSpaces = 0,
-        DoNotIgnoreWhiteSpaces = 1,
-    }
-    public sealed partial class Oid
-    {
-        public Oid() { }
-        public Oid(System.Security.Cryptography.Oid oid) { }
-        public Oid(string oid) { }
-        public Oid(string? value, string? friendlyName) { }
-        public string? FriendlyName { get { throw null; } set { } }
-        public string? Value { get { throw null; } set { } }
-        public static System.Security.Cryptography.Oid FromFriendlyName(string friendlyName, System.Security.Cryptography.OidGroup group) { throw null; }
-        public static System.Security.Cryptography.Oid FromOidValue(string oidValue, System.Security.Cryptography.OidGroup group) { throw null; }
-    }
-    public sealed partial class OidCollection : System.Collections.ICollection, System.Collections.IEnumerable
-    {
-        public OidCollection() { }
-        public int Count { get { throw null; } }
-        public bool IsSynchronized { get { throw null; } }
-        public System.Security.Cryptography.Oid this[int index] { get { throw null; } }
-        public System.Security.Cryptography.Oid? this[string oid] { get { throw null; } }
-        public object SyncRoot { get { throw null; } }
-        public int Add(System.Security.Cryptography.Oid oid) { throw null; }
-        public void CopyTo(System.Security.Cryptography.Oid[] array, int index) { }
-        public System.Security.Cryptography.OidEnumerator GetEnumerator() { throw null; }
-        void System.Collections.ICollection.CopyTo(System.Array array, int index) { }
-        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() { throw null; }
-    }
-    public sealed partial class OidEnumerator : System.Collections.IEnumerator
-    {
-        internal OidEnumerator() { }
-        public System.Security.Cryptography.Oid Current { get { throw null; } }
-        object System.Collections.IEnumerator.Current { get { throw null; } }
-        public bool MoveNext() { throw null; }
-        public void Reset() { }
-    }
-    public enum OidGroup
-    {
-        All = 0,
-        HashAlgorithm = 1,
-        EncryptionAlgorithm = 2,
-        PublicKeyAlgorithm = 3,
-        SignatureAlgorithm = 4,
-        Attribute = 5,
-        ExtensionOrAttribute = 6,
-        EnhancedKeyUsage = 7,
-        Policy = 8,
-        Template = 9,
-        KeyDerivationFunction = 10,
-    }
-    public static partial class PemEncoding
-    {
-        public static System.Security.Cryptography.PemFields Find(System.ReadOnlySpan<char> pemData) { throw null; }
-        public static int GetEncodedSize(int labelLength, int dataLength) { throw null; }
-        public static bool TryFind(System.ReadOnlySpan<char> pemData, out System.Security.Cryptography.PemFields fields) { throw null; }
-        public static bool TryWrite(System.ReadOnlySpan<char> label, System.ReadOnlySpan<byte> data, System.Span<char> destination, out int charsWritten) { throw null; }
-        public static char[] Write(System.ReadOnlySpan<char> label, System.ReadOnlySpan<byte> data) { throw null; }
-    }
-    public readonly partial struct PemFields
-    {
-        private readonly object _dummy;
-        private readonly int _dummyPrimitive;
-        public System.Range Base64Data { get { throw null; } }
-        public int DecodedDataLength { get { throw null; } }
-        public System.Range Label { get { throw null; } }
-        public System.Range Location { get { throw null; } }
-    }
-    public partial class ToBase64Transform : System.IDisposable, System.Security.Cryptography.ICryptoTransform
-    {
-        public ToBase64Transform() { }
-        public virtual bool CanReuseTransform { get { throw null; } }
-        public bool CanTransformMultipleBlocks { get { throw null; } }
-        public int InputBlockSize { get { throw null; } }
-        public int OutputBlockSize { get { throw null; } }
-        public void Clear() { }
-        public void Dispose() { }
-        protected virtual void Dispose(bool disposing) { }
-        ~ToBase64Transform() { }
-        public int TransformBlock(byte[] inputBuffer, int inputOffset, int inputCount, byte[] outputBuffer, int outputOffset) { throw null; }
-        public byte[] TransformFinalBlock(byte[] inputBuffer, int inputOffset, int inputCount) { throw null; }
-    }
-}
