+++ src/libraries/System.Security.Cryptography.Encoding/ref/System.Security.Cryptography.Encoding.csproj
@@ -4,7 +4,7 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="System.Security.Cryptography.Encoding.cs" />
+    <Compile Include="System.Security.Cryptography.Encoding.Forwards.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
