--- src/libraries/System.Security.Cryptography.Encoding/src/Resources/Strings.resx
@@ -1,144 +1,0 @@
-﻿<root>
-  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
-    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
-    <xsd:element name="root" msdata:IsDataSet="true">
-      <xsd:complexType>
-        <xsd:choice maxOccurs="unbounded">
-          <xsd:element name="metadata">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" />
-              </xsd:sequence>
-              <xsd:attribute name="name" use="required" type="xsd:string" />
-              <xsd:attribute name="type" type="xsd:string" />
-              <xsd:attribute name="mimetype" type="xsd:string" />
-              <xsd:attribute ref="xml:space" />
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="assembly">
-            <xsd:complexType>
-              <xsd:attribute name="alias" type="xsd:string" />
-              <xsd:attribute name="name" type="xsd:string" />
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="data">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
-                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
-              </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
-              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
-              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
-              <xsd:attribute ref="xml:space" />
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="resheader">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
-              </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" />
-            </xsd:complexType>
-          </xsd:element>
-        </xsd:choice>
-      </xsd:complexType>
-    </xsd:element>
-  </xsd:schema>
-  <resheader name="resmimetype">
-    <value>text/microsoft-resx</value>
-  </resheader>
-  <resheader name="version">
-    <value>2.0</value>
-  </resheader>
-  <resheader name="reader">
-    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <resheader name="writer">
-    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <data name="Cryptography_Oid_InvalidValue" xml:space="preserve">
-    <value>The OID value is invalid.</value>
-  </data>
-  <data name="Cryptography_Oid_InvalidName" xml:space="preserve">
-    <value>No OID value matches this name.</value>
-  </data>
-  <data name="Cryptography_Oid_SetOnceValue" xml:space="preserve">
-    <value>The OID value cannot be changed from its current value.</value>
-  </data>
-  <data name="Cryptography_Oid_SetOnceFriendlyName" xml:space="preserve">
-    <value>The OID friendly name cannot be changed from its current value.</value>
-  </data>
-  <data name="Cryptography_SSE_InvalidDataSize" xml:space="preserve">
-    <value>NoLength of the data to encrypt is invalid.</value>
-  </data>
-  <data name="Cryptography_Asn_EnumeratedValueRequiresNonFlagsEnum" xml:space="preserve">
-    <value>ASN.1 Enumerated values only apply to enum types without the [Flags] attribute.</value>
-  </data>
-  <data name="Cryptography_Asn_NamedBitListRequiresFlagsEnum" xml:space="preserve">
-    <value>Named bit list operations require an enum with the [Flags] attribute.</value>
-  </data>
-  <data name="Cryptography_Asn_NamedBitListValueTooBig" xml:space="preserve">
-    <value>The encoded named bit list value is larger than the value size of the '{0}' enum.</value>
-  </data>
-  <data name="Cryptography_Asn_UniversalValueIsFixed" xml:space="preserve">
-    <value>Tags with TagClass Universal must have the appropriate TagValue value for the data type being read or written.</value>
-  </data>
-  <data name="Cryptography_Asn_UnusedBitCountRange" xml:space="preserve">
-    <value>Unused bit count must be between 0 and 7, inclusive.</value>
-  </data>
-  <data name="Cryptography_AsnWriter_EncodeUnbalancedStack" xml:space="preserve">
-    <value>Encode cannot be called while a Sequence or SetOf is still open.</value>
-  </data>
-  <data name="Cryptography_AsnWriter_PopWrongTag" xml:space="preserve">
-    <value>Cannot pop the requested tag as it is not currently in progress.</value>
-  </data>
-  <data name="Cryptography_Der_Invalid_Encoding" xml:space="preserve">
-    <value>ASN1 corrupted data.</value>
-  </data>
-  <data name="Cryptography_Invalid_IA5String" xml:space="preserve">
-    <value>The string contains a character not in the 7 bit ASCII character set.</value>
-  </data>
-  <data name="Cryptography_WriteEncodedValue_OneValueAtATime" xml:space="preserve">
-    <value>The input to WriteEncodedValue must represent a single encoded value with no trailing data.</value>
-  </data>
-  <data name="Arg_RankMultiDimNotSupported" xml:space="preserve">
-    <value>Only single dimensional arrays are supported for the requested action.</value>
-  </data>
-  <data name="ArgumentOutOfRange_Index" xml:space="preserve">
-    <value>Index was out of range. Must be non-negative and less than the size of the collection.</value>
-  </data>
-  <data name="ArgumentOutOfRange_NeedNonNegNum" xml:space="preserve">
-    <value>Non-negative number required.</value>
-  </data>
-  <data name="Argument_EncodeDestinationTooSmall" xml:space="preserve">
-    <value>The destination is too small to hold the encoded value.</value>
-  </data>
-  <data name="Argument_InvalidOidValue" xml:space="preserve">
-    <value>The OID value was invalid.</value>
-  </data>
-  <data name="Argument_InvalidOffLen" xml:space="preserve">
-    <value>Offset and length were out of bounds for the array or count is greater than the number of elements from index to the end of the source collection.</value>
-  </data>
-  <data name="Argument_InvalidValue" xml:space="preserve">
-    <value>Value was invalid.</value>
-  </data>
-  <data name="Argument_PemEncoding_NoPemFound" xml:space="preserve">
-    <value>No PEM encoded data found.</value>
-  </data>
-  <data name="Argument_PemEncoding_InvalidLabel" xml:space="preserve">
-    <value>The specified label is not valid.</value>
-  </data>
-  <data name="Argument_PemEncoding_EncodedSizeTooLarge" xml:space="preserve">
-    <value>The encoded PEM size is too large to represent as a signed 32-bit integer.</value>
-  </data>
-  <data name="ArgumentOutOfRange_NeedPositiveNumber" xml:space="preserve">
-    <value>A positive number is required.</value>
-  </data>
-  <data name="ObjectDisposed_Generic" xml:space="preserve">
-    <value>Cannot access a disposed object.</value>
-  </data>
-  <data name="SystemSecurityCryptographyEncoding_PlatformNotSupported" xml:space="preserve">
-    <value>System.Security.Cryptography.Encoding is not supported on this platform.</value>
-  </data>
-</root>
