--- src/libraries/System.Security.Cryptography.Encoding/tests/System.Security.Cryptography.Encoding.Tests.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <TargetFrameworks>$(NetCoreAppCurrent)-windows;$(NetCoreAppCurrent)-Unix;$(NetCoreAppCurrent)-Browser</TargetFrameworks>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="AsnEncodedData.cs" />
-    <Compile Include="AsnEncodedDataCollectionTests.cs" />
-    <Compile Include="Base64TransformsTests.cs" />
-    <Compile Include="Oid.cs" />
-    <Compile Include="OidCollectionTests.cs" />
-    <Compile Include="PemEncodingFindTests.cs" />
-    <Compile Include="PemEncodingTests.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\ByteUtils.cs" Link="CommonTest\System\Security\Cryptography\ByteUtils.cs" />
-    <Compile Include="$(CommonTestPath)System\Security\Cryptography\PlatformSupport.cs" Link="CommonTest\System\Security\Cryptography\PlatformSupport.cs" />
-    <Compile Include="$(CommonPath)System\Memory\PointerMemoryManager.cs" Link="Common\System\Memory\PointerMemoryManager.cs" />
-  </ItemGroup>
-</Project>
