+++ src/libraries/System.Security.Cryptography/src/Resources/Strings.resx
@@ -60,6 +60,9 @@
   <data name="Arg_CryptographyException" xml:space="preserve">
     <value>Error occurred during a cryptographic operation.</value>
   </data>
+  <data name="Arg_RankMultiDimNotSupported" xml:space="preserve">
+    <value>Only single dimensional arrays are supported for the requested action.</value>
+  </data>
   <data name="Argument_DestinationTooShort" xml:space="preserve">
     <value>Destination is too short.</value>
   </data>
@@ -78,6 +81,18 @@
   <data name="Argument_BitsMustBeWholeBytes" xml:space="preserve">
     <value>The value specified in bits must be a whole number of bytes.</value>
   </data>
+  <data name="Argument_PemEncoding_NoPemFound" xml:space="preserve">
+    <value>No PEM encoded data found.</value>
+  </data>
+  <data name="Argument_PemEncoding_InvalidLabel" xml:space="preserve">
+    <value>The specified label is not valid.</value>
+  </data>
+  <data name="Argument_PemEncoding_EncodedSizeTooLarge" xml:space="preserve">
+    <value>The encoded PEM size is too large to represent as a signed 32-bit integer.</value>
+  </data>
+  <data name="ArgumentOutOfRange_Index" xml:space="preserve">
+    <value>Index was out of range. Must be non-negative and less than the size of the collection.</value>
+  </data>
   <data name="ArgumentOutOfRange_NeedNonNegNum" xml:space="preserve">
     <value>Non-negative number required.</value>
   </data>
@@ -89,6 +104,9 @@
   </data>
   <data name="Cryptography_DefaultAlgorithm_NotSupported" xml:space="preserve">
     <value>This platform does not allow the automatic selection of an algorithm.</value>
+  </data>
+  <data name="Cryptography_Der_Invalid_Encoding" xml:space="preserve">
+    <value>ASN1 corrupted data.</value>
   </data>
   <data name="Cryptography_HashNotYetFinalized" xml:space="preserve">
     <value>Hash must be finalized before the hash value is retrieved.</value>
@@ -120,6 +138,18 @@
   <data name="Cryptography_MatchFeedbackSize" xml:space="preserve">
     <value>The specified plaintext size is not valid for the the padding and feedback size.</value>
   </data>
+  <data name="Cryptography_Oid_InvalidName" xml:space="preserve">
+    <value>No OID value matches this name.</value>
+  </data>
+  <data name="Cryptography_Oid_InvalidValue" xml:space="preserve">
+    <value>The OID value is invalid.</value>
+  </data>
+  <data name="Cryptography_Oid_SetOnceFriendlyName" xml:space="preserve">
+    <value>The OID friendly name cannot be changed from its current value.</value>
+  </data>
+  <data name="Cryptography_Oid_SetOnceValue" xml:space="preserve">
+    <value>The OID value cannot be changed from its current value.</value>
+  </data>
   <data name="Cryptography_PlaintextTooLarge" xml:space="preserve">
     <value>The specified plaintext size is too large.</value>
   </data>
@@ -149,5 +179,8 @@
   </data>
   <data name="InvalidOperation_UnsupportedBlockSize" xml:space="preserve">
     <value>The algorithm's block size is not supported.</value>
+  </data>
+  <data name="SystemSecurityCryptography_PlatformNotSupported" xml:space="preserve">
+    <value>System.Security.Cryptography is not supported on this platform.</value>
   </data>
 </root>
