+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/AsnFormatter.Managed.cs
@@ -1,13 +1,11 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using System.Formats.Asn1;
-using System.Security.Cryptography;
 using System.Security.Cryptography.Asn1;
 using System.Text;
 
-namespace Internal.Cryptography
+namespace System.Security.Cryptography
 {
     internal abstract partial class AsnFormatter
     {
