+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/AsnFormatter.Windows.cs
@@ -1,10 +1,10 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Internal.Cryptography
+namespace System.Security.Cryptography
 {
     internal abstract partial class AsnFormatter
     {
-        private static readonly AsnFormatter s_instance = new OpenSslAsnFormatter();
+        private static readonly AsnFormatter s_instance = new CngAsnFormatter();
     }
 }
