+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/CngAsnFormatter.cs
@@ -1,14 +1,10 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using System.Buffers;
 using System.Runtime.InteropServices;
-using System.Security.Cryptography;
 
-using Internal.NativeCrypto;
-
-namespace Internal.Cryptography
+namespace System.Security.Cryptography
 {
     internal sealed class CngAsnFormatter : AsnFormatter
     {
