+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/OidLookup.Windows.cs
@@ -1,11 +1,9 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Security.Cryptography;
-
 using static Interop.Crypt32;
 
-namespace Internal.Cryptography
+namespace System.Security.Cryptography
 {
     internal static partial class OidLookup
     {
