+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/OpenSslAsnFormatter.cs
@@ -1,12 +1,11 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Security.Cryptography;
 using System.Text;
 
 using Microsoft.Win32.SafeHandles;
 
-namespace Internal.Cryptography
+namespace System.Security.Cryptography
 {
     internal sealed class OpenSslAsnFormatter : AsnFormatter
     {
