+++ src/libraries/System.Security.Cryptography/tests/AsnEncodedDataCollectionTests.cs
@@ -1,14 +1,12 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.IO;
-using System.Text;
 using System.Linq;
 using System.Collections;
 using System.Collections.Generic;
 using Xunit;
 
-namespace System.Security.Cryptography.Encoding.Tests
+namespace System.Security.Cryptography.Tests
 {
     public static class AsnEncodedDataCollectionTests
     {
