+++ src/libraries/System.Security.Cryptography/tests/Base64TransformsTests.cs
@@ -6,7 +6,7 @@ using System.IO;
 using System.Linq;
 using Xunit;
 
-namespace System.Security.Cryptography.Encoding.Tests
+namespace System.Security.Cryptography.Tests
 {
     public class Base64TransformsTests
     {
