+++ src/libraries/System.Security.Cryptography/tests/OidTests.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using Test.Cryptography;
 using Xunit;
 
-namespace System.Security.Cryptography.Encoding.Tests
+namespace System.Security.Cryptography.Tests
 {
     public static class OidTests
     {
