+++ src/libraries/System.Security.Cryptography/tests/PemEncodingFindTests.cs
@@ -1,12 +1,10 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Collections.Generic;
-using System.Security.Cryptography;
 using System.Text;
 using Xunit;
 
-namespace System.Security.Cryptography.Encoding.Tests
+namespace System.Security.Cryptography.Tests
 {
     public abstract class PemEncodingFindTests
     {
